Change login to be session based.
project3

+++ public/project/server/services/user.service.server.js
@@ -3,16 +3,29 @@ module.exports = function(app, UserModel) {
 
     var utils = require('./util.js')();
 
-    var baseUrl = '/api/project/user';
-    var idParam = baseUrl + '/:id';
+    var BASE_URL = '/api/project/user';
+    var ID_PARAM_URL = BASE_URL + '/:id';
+    
+    var LOGIN_URL = '/api/project/login';
+    var LOGGED_IN_URL = '/api/project/loggedIn';
+    var LOGOUT_URL = '/api/project/logout';
+
     var errorMsg = 'User not found';
 
-    app.post(baseUrl, createUser);
-    app.get(baseUrl, getUser);
-    app.get(idParam, getUserById);
-    app.put(idParam, updateUser);
-    app.delete(idParam, deleteUser);
+    app.post(BASE_URL, createUser);
+    app.get(BASE_URL, getUser);
+   
+    app.get(ID_PARAM_URL, getUserById);
+    app.put(ID_PARAM_URL, updateUser);
+   
+    app.delete(ID_PARAM_URL, deleteUser);
 
+    app.post(LOGIN_URL, login);
+    app.get(LOGGED_IN_URL, loggedIn);
+    app.post(LOGOUT_URL, logout);
+
+    /////////////////////////////////////////
+    
     function createUser(req, res) {
         var user = UserModel.create(req.body);
         
@@ -27,9 +40,7 @@ module.exports = function(app, UserModel) {
         var username = req.query.username;
         var password = req.query.password;
 
-        if (username && password) {
-            getUserByCredentials({ username: username, password: password }, res);
-        } else if (username) {
+        if (username) {
             getUserByUsername(username, res);
         } else {
             getAllUsers(res);
@@ -43,12 +54,47 @@ module.exports = function(app, UserModel) {
 
     function updateUser(req, res) {
         var id = req.params.id;
-        utils.sendOr404(UserModel.update(id, req.body), res, errorMsg);
+
+        if (req.session.currentUser) {
+            res.status(401).send('Not authorized');
+        } else if (req.session.currentUser.id !== id) {
+            res.status(403).send('You cannot update another user.');
+        } else {
+            utils.sendOr404(UserModel.update(id, req.body), res, errorMsg);
+        }
     }
 
     function deleteUser(req, res) {
         var id = req.params.id;
-        utils.sendOr404(UserModel.delete(id), res, errorMsg);
+        
+        if (req.session.currentUser) {
+            res.status(401).send('Not authorized');
+        } else if (req.session.currentUser.id !== id) {
+            res.status(403).send('You cannot update another user.');
+        } else {
+            utils.sendOr404(UserModel.delete(id), res, errorMsg);
+        }
+    }
+
+    function login(req, res) {
+        var credentials = req.body;
+        var user = UserModel.findByCredentials(credentials);
+
+        if (user) {
+            req.session.currentUser = user;
+            res.json(user);
+        } else {
+            res.status(404).send(errorMsg);
+        }
+    }
+
+    function loggedIn(req, res) {
+        res.json(req.session.currentUser);
+    }
+
+    function logout(req, res) {
+        req.session.destroy();
+        res.send(200);
     }
 
     function getUserByCredentials(credentials, res) {
